sessionctx/variable: small refactor (split large file) (#30511)
